Rename DistributedContext to CorrelationContext in the API/SDK.
