+++ sdk/src/main/resources/META-INF/services/io.opentelemetry.correlationcontext.spi.CorrelationContextManagerProvider
@@ -1,0 +1,1 @@
+io.opentelemetry.sdk.correlationcontext.CorrelationContextManagerSdkProvider
