--- sdk/src/main/resources/META-INF/services/io.opentelemetry.distributedcontext.spi.DistributedContextManagerProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.sdk.distributedcontext.DistributedContextManagerSdkProvider
