+++ sdk/src/test/java/io/opentelemetry/sdk/correlationcontext/CorrelationContextManagerSdkProviderTest.java
@@ -14,25 +14,20 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.sdk;
+package io.opentelemetry.sdk.correlationcontext;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opentelemetry.OpenTelemetry;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
-public class OpenTelemetrySdkTest {
+public class CorrelationContextManagerSdkProviderTest {
 
   @Test
   public void testDefault() {
-    assertThat(OpenTelemetrySdk.getTracerRegistry())
-        .isSameInstanceAs(OpenTelemetry.getTracerRegistry());
-    assertThat(OpenTelemetrySdk.getCorrelationContextManager())
-        .isSameInstanceAs(OpenTelemetry.getCorrelationContextManager());
-    assertThat(OpenTelemetrySdk.getMeterRegistry())
-        .isSameInstanceAs(OpenTelemetry.getMeterRegistry());
+    assertThat(new CorrelationContextManagerSdkProvider().create())
+        .isInstanceOf(CorrelationContextManagerSdk.class);
   }
 }
