+++ sdk/src/test/java/io/opentelemetry/sdk/correlationcontext/CorrelationContextTestUtil.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2019, OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.opentelemetry.sdk.correlationcontext;
+
+import io.opentelemetry.correlationcontext.Entry;
+import java.util.Arrays;
+import java.util.List;
+
+class CorrelationContextTestUtil {
+
+  static CorrelationContextSdk listToCorrelationContext(Entry... entries) {
+    return listToCorrelationContext(Arrays.asList(entries));
+  }
+
+  static CorrelationContextSdk listToCorrelationContext(List<Entry> entries) {
+    CorrelationContextSdk.Builder builder = new CorrelationContextSdk.Builder();
+    for (Entry entry : entries) {
+      builder.put(entry.getKey(), entry.getValue(), entry.getEntryMetadata());
+    }
+    return builder.build();
+  }
+
+  private CorrelationContextTestUtil() {}
+}
