--- sdk/src/test/java/io/opentelemetry/sdk/distributedcontext/DistributedContextTestUtil.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2019, OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.opentelemetry.sdk.distributedcontext;
-
-import io.opentelemetry.distributedcontext.Entry;
-import java.util.Arrays;
-import java.util.List;
-
-class DistributedContextTestUtil {
-
-  static DistributedContextSdk listToDistributedContext(Entry... entries) {
-    return listToDistributedContext(Arrays.asList(entries));
-  }
-
-  static DistributedContextSdk listToDistributedContext(List<Entry> entries) {
-    DistributedContextSdk.Builder builder = new DistributedContextSdk.Builder();
-    for (Entry entry : entries) {
-      builder.put(entry.getKey(), entry.getValue(), entry.getEntryMetadata());
-    }
-    return builder.build();
-  }
-
-  private DistributedContextTestUtil() {}
-}
